feat:#3 implements PokemonDataSourceImp

+++ app/src/main/java/com/example/pokemondex/data/PokemonModel.kt
@@ -3,7 +3,7 @@ package com.example.pokemondex.data
 data class PokemonModel(
     val id: Int,
     val name: String,
-    val height: Float,
-    val weight: Float,
+    val height: Int,
+    val weight: Int,
     val pokeImageUrl: String
 )
